Changed Integer to Long for Numeric value

+++ src/Procon.Database.Serialization.Test/TestSerializerFind.cs
@@ -209,7 +209,7 @@ namespace Procon.Database.Serialization.Test {
                         Name = "Score"
                     },
                     new NumericValue() {
-                        Integer = 10
+                        Long = 10
                     }
                 })
             .Collection(new Collection() {
@@ -284,7 +284,7 @@ namespace Procon.Database.Serialization.Test {
                     Name = "Score"
                 },
                 new NumericValue() {
-                    Integer = 10
+                    Long = 10
                 }
             })
             .Condition(new LessThan() {
@@ -292,7 +292,7 @@ namespace Procon.Database.Serialization.Test {
                     Name = "Score"
                 },
                 new NumericValue() {
-                    Integer = 20
+                    Long = 20
                 }
             })
             .Collection(new Collection() {
@@ -327,7 +327,7 @@ namespace Procon.Database.Serialization.Test {
                                 Name = "Score"
                             },
                             new NumericValue() {
-                                Integer = 50
+                                Long = 50
                             }
                         }
                     },
@@ -345,7 +345,7 @@ namespace Procon.Database.Serialization.Test {
                                 Name = "Score"
                             },
                             new NumericValue() {
-                                Integer = 50
+                                Long = 50
                             }
                         }
                     }
